sentris_rag/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentris_rag"
version = "0.1.0"
description = "Educational RAG system for Sentris platform"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["education", "rag", "llm", "deepseek", "nlp"]
authors = [{ name = "Sentris Team" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "aiohttp>=3.8.0",
    "beautifulsoup4>=4.9.3",
    "chromadb>=0.4.0",
    "duckduckgo-search>=3.0.0",
    "google-api-python-client>=2.0.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "pydantic>=2.0.0",
    "python-docx>=0.8.11",
    "PyPDF2>=3.0.0",
    "sentence-transformers>=2.2.0",
    "tiktoken>=0.3.0",
    "toml>=0.10.2",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.3.0",
    "isort>=5.10.1",
    "mypy>=0.981",
    "pytest>=7.1.3",
    "pytest-asyncio>=0.19.0",
]

[project.urls]
Homepage = "https://github.com/sentris/sentris_rag"
Repository = "https://github.com/sentris/sentris_rag.git"
Documentation = "https://sentris.github.io/sentris_rag"

[tool.hatch.build.targets.wheel]
packages = ["src/sentris_rag"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
select = ["E", "F", "B"]
ignore = []
line-length = 88
target-version = "py38"
